Ignore fake empty selection responses
For some reason, on some pages the content script side not only
seems to think there is no selection, but sends several responses.
So it would send the selection once, among many empty responses. This
would make the extension fail, since it only takes into account the
first response (which would be empty in that case).

This change makes sure the content script only sends a response if/when
it found a non empty selection.