googdoc: another chrome extension, but for Google Docs
The regular extension cannot access the contents of a Google Docs with
the naive way. So this new extension uses
github.com/JensPLarsen/ChromeExtension-GoogleDocsUtil to access a Google
Docs contents, and in particular the highlighted text.

I might merge the two later, we'll see.

Unfortunately, GoogleDocsUtil does not enable to modify the contents of
the doc, so I'm still displaying the translation in the extension popup
for now. It's kinda lame but maybe that will work for now.
Maybe it wouldn't be so hard to extend GoogleDocsUtil to do writes in
addition to reads?